# docs/swagger.yaml
info:
  contact: {}
paths:
  /admin/category-creat:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ' name'
        in: formData
        name: name
        required: true
        type: string
      - description: parent_id
        in: formData
        name: parent_id
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 创建分类
      tags:
      - 管理员私有方法
  /admin/category-delete:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: identity
        in: query
        name: identity
        required: true
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 分类删除
      tags:
      - 管理员私有方法
  /admin/category-list:
    get:
      parameters:
      - description: 请输入当前页，默认第一页
        in: query
        name: page
        type: integer
      - description: 请输入每页数量，默认20条
        in: query
        name: size
        type: integer
      - description: 请输入类别名称
        in: query
        name: name
        type: string
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 获取分类列表
      tags:
      - 管理员私有方法
  /admin/category-modify:
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: identity
        in: formData
        name: identity
        required: true
        type: string
      - description: ' name'
        in: formData
        name: name
        required: true
        type: string
      - description: parent_id
        in: formData
        name: parent_id
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 分类修改
      tags:
      - 管理员私有方法
  /admin/problem-create:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: title
        in: formData
        name: title
        required: true
        type: string
      - description: content
        in: formData
        name: content
        required: true
        type: string
      - description: max_runtime
        in: formData
        name: max_runtime
        type: integer
      - description: max_memory
        in: formData
        name: max_memory
        type: integer
      - collectionFormat: multi
        description: category_ids
        in: formData
        items:
          type: string
        name: category_ids
        type: array
      - collectionFormat: multi
        description: test_cases
        in: formData
        items:
          type: string
        name: test_cases
        type: array
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 问题创建
      tags:
      - 管理员私有方法
  /admin/problem-modify:
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: identity
        in: formData
        name: identity
        required: true
        type: string
      - description: title
        in: formData
        name: title
        required: true
        type: string
      - description: ' content'
        in: formData
        name: content
        required: true
        type: string
      - collectionFormat: multi
        description: category_ids
        in: formData
        items:
          type: string
        name: category_ids
        type: array
      - collectionFormat: multi
        description: test_cases
        in: formData
        items:
          type: string
        name: test_cases
        type: array
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 问题修改
      tags:
      - 管理员私有方法
  /admin/submit-code:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: problem_identity
        in: formData
        name: problem_identity
        required: true
        type: string
      - description: code
        in: body
        name: code
        required: true
        schema:
          type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 提交代码
      tags:
      - 用户私有方法
  /login:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - 公共方法
  /problem-detail:
    get:
      parameters:
      - description: problem identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 问题详情
      tags:
      - 公共方法
  /problem-list:
    get:
      parameters:
      - description: 请输入当前页，默认第一页
        in: query
        name: page
        type: integer
      - description: 请输入每页数量，默认10条
        in: query
        name: size
        type: integer
      - description: 请输入关键字
        in: query
        name: keyword
        type: string
      - description: 请输入分类id
        in: query
        name: category_identity
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 问题列表
      tags:
      - 公共方法
  /register:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      - description: code
        in: formData
        name: code
        required: true
        type: string
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 用户注册
      tags:
      - 公共方法
  /send-code:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 发送验证码
      tags:
      - 公共方法
  /submit-list:
    get:
      parameters:
      - description: user_identity
        in: query
        name: user_identity
        type: string
      - description: problem_identity
        in: query
        name: problem_identity
        type: string
      - description: status
        in: query
        name: status
        type: integer
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 提交表详情
      tags:
      - 公共方法
  /user-detail:
    get:
      parameters:
      - description: user identity
        in: query
        name: identity
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 用户详情
      tags:
      - 公共方法
  /user-rank:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      responses:
        "200":
          description: '{"code":200,"msg":"success","data":""}'
          schema:
            type: string
      summary: 用户排名
      tags:
      - 公共方法
swagger: "2.0"